Core/Spells: Extended spell required shapeshift masks to 64 bits and defined a new spell attribute
(cherry picked from commit 191d2c7262d3680776579055a063a433add54277)

+++ src/server/game/Spells/SpellInfo.h
@@ -306,8 +306,8 @@ public:
     uint32 AttributesEx6;
     uint32 AttributesEx7;
     uint32 AttributesCu;
-    uint32 Stances;
-    uint32 StancesNot;
+    uint64 Stances;
+    uint64 StancesNot;
     uint32 Targets;
     uint32 TargetCreatureType;
     uint32 RequiresSpellFocus;
